work on 'in progress' trip

+++ web/src/main/java/com/inthinc/pro/backing/BasePerformanceBean.java
@@ -110,7 +110,7 @@ public abstract class BasePerformanceBean extends BaseBean
         {
             // Get mileage for day.
             if (dailyList.get(i).getIdentifierNum() != null)
-                odometerValues[i] = MeasurementConversionUtil.convertMilesToKilometers((dailyList.get(i).getIdentifierNum()  / 100D), getPerson().getMeasurementType()).doubleValue();
+                odometerValues[i] = MeasurementConversionUtil.convertMilesToKilometers((dailyList.get(i).getIdentifierNum().longValue()  / 100D), getPerson().getMeasurementType()).doubleValue();
 
             // Set Score to NULL on non driving days.
             if (odometerValues[i] == null || odometerValues[i] == 0)
